decoders. add elevation, fix point attribute calculation

+++ nebula_decoders/include/nebula_decoders/nebula_decoders_velodyne/decoders/velodyne_scan_decoder.hpp
@@ -39,7 +39,7 @@ static const int RAW_SCAN_SIZE = 3;
 static const int SCANS_PER_BLOCK = 32;
 static const int BLOCK_DATA_SIZE = (SCANS_PER_BLOCK * RAW_SCAN_SIZE);
 
-static const float ROTATION_RESOLUTION = 0.01f;     // [deg]
+static const double ROTATION_RESOLUTION = 0.01;     // [deg]
 static const uint16_t ROTATION_MAX_UNITS = 36000u;  // [deg/100]
 
 static const size_t RETURN_MODE_INDEX = 1204;
